IGNITE-14971 Thin client: Add RecordView (#497)
* Implement `ClientRecordView` with POJO support.
  * 1-to-1 field name to column mapping.
  * get and put operations only to limit the size of PR (to be continued in [IGNITE-16087](https://issues.apache.org/jira/browse/IGNITE-16087)).
* Extract common serialization logic from `ignite-schema` into `ignite-marshaller-common` module (proper code reuse to be achieved later: [IGNITE-16088](https://issues.apache.org/jira/browse/IGNITE-16088)).

+++ modules/client-common/src/main/java/org/apache/ignite/internal/client/proto/ClientMessageUnpacker.java
@@ -342,7 +342,7 @@ public class ClientMessageUnpacker implements AutoCloseable {
                 return BigInteger.valueOf(buf.readLong());
 
             default:
-                throw unexpected("Integer", code);
+                throw unexpected("BigInteger", code);
         }
     }
